pyproject.toml
[project]
name = "gpt-batcher"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
]
